+++ tests/core/RssResponseTransformationTests.cpp
@@ -65,6 +65,12 @@ TEST_F(RssResponseTransformationTests, invalidTimeStamp)
 
   ASSERT_FALSE(::ad_rss::core::RssResponseTransformation::transformProperResponse(
     worldModel, responseState, accelerationRestriction));
+
+  worldModel.timeIndex = 0u;
+  responseState.timeIndex = 1u;
+
+  ASSERT_FALSE(::ad_rss::core::RssResponseTransformation::transformProperResponse(
+    worldModel, responseState, accelerationRestriction));
 }
 
 TEST_F(RssResponseTransformationTests, invalidLongitudinalResponseState)