+++ tests/core/RssSituationExtractionSameDirectionTests.cpp
@@ -104,7 +104,7 @@ TEST_F(RssSituationExtractionSameDirectionTests, noLongitudinalDifference)
 
   scene.egoVehicleRoad.push_back(roadSegment);
   worldModel.scenes.push_back(scene);
-  worldModel.timeIndex = 0;
+  worldModel.timeIndex = 1;
 
   ASSERT_TRUE(extractSituations(worldModel, situationVector));
   ASSERT_EQ(situationVector.size(), 1);
@@ -145,7 +145,7 @@ TEST_F(RssSituationExtractionSameDirectionTests, longitudinalDifferenceEgoLeadin
 
   scene.egoVehicleRoad.push_back(roadSegment);
   worldModel.scenes.push_back(scene);
-  worldModel.timeIndex = 0;
+  worldModel.timeIndex = 1;
 
   ASSERT_TRUE(extractSituations(worldModel, situationVector));
   ASSERT_EQ(situationVector.size(), 1);
@@ -185,7 +185,7 @@ TEST_F(RssSituationExtractionSameDirectionTests, longitudinalDifferenceEgoFollow
 
   scene.egoVehicleRoad.push_back(roadSegment);
   worldModel.scenes.push_back(scene);
-  worldModel.timeIndex = 0;
+  worldModel.timeIndex = 1;
 
   ASSERT_TRUE(extractSituations(worldModel, situationVector));
   ASSERT_EQ(situationVector.size(), 1);